requirements/base.txt
django==1.10.7
django-environ==0.4.4
psycopg2==2.7.1
unidecode==0.4.21
Pillow==4.3.0
django-dbbackup==3.2.0